--- src/Workflow/Hydrator/WorkflowHydratorService.php
@@ -1,71 +1,0 @@
-<?php
-declare(strict_types=1);
-
-/**
- * Pimcore
- *
- * This source file is available under following license:
- * - Pimcore Commercial License (PCL)
- *
- * @copyright  Copyright (c) Pimcore GmbH (http://www.pimcore.org)
- * @license    http://www.pimcore.org/license     PCL
- */
-
-namespace Pimcore\Bundle\StudioBackendBundle\Workflow\Hydrator;
-
-use Pimcore\Bundle\StaticResolverBundle\Models\Element\ServiceResolverInterface;
-use Pimcore\Bundle\StudioBackendBundle\Security\Service\SecurityServiceInterface;
-use Pimcore\Bundle\StudioBackendBundle\Util\Constants\ElementPermissions;
-use Pimcore\Bundle\StudioBackendBundle\Util\Traits\ElementProviderTrait;
-use Pimcore\Bundle\StudioBackendBundle\Workflow\Request\WorkflowDetailsParameters;
-use Pimcore\Bundle\StudioBackendBundle\Workflow\Schema\WorkflowDetails;
-use Pimcore\Model\UserInterface;
-use Pimcore\Workflow\Manager;
-
-/**
- * @internal
- */
-final readonly class WorkflowHydratorService implements WorkflowHydratorServiceInterface
-{
-    use ElementProviderTrait;
-
-    public function __construct(
-        private Manager $workflowManager,
-        private SecurityServiceInterface $securityService,
-        private ServiceResolverInterface $serviceResolver,
-        private WorkflowDetailsHydratorInterface $workflowDetailsHydrator,
-    ) {
-    }
-
-    /**
-     * @return WorkflowDetails[]
-     */
-    public function hydrateWorkflowDetails(
-        WorkflowDetailsParameters $parameters,
-        UserInterface $user
-    ): array
-    {
-        $element = $this->getElement(
-            $this->serviceResolver,
-            $parameters->getElementType(),
-            $parameters->getElementId(),
-        );
-
-        $this->securityService->hasElementPermission(
-            $element,
-            $user,
-            ElementPermissions::VIEW_PERMISSION
-        );
-
-        $details =  [];
-        $elementWorkflows = $this->workflowManager->getAllWorkflowsForSubject($element);
-        foreach ($elementWorkflows as $workflow) {
-            $details[] = $this->workflowDetailsHydrator->hydrate(
-                $element,
-                $workflow
-            );
-        }
-
-        return $details;
-    }
-}
